rename til mer forstålig klassenavn fra gdpr til audit
