+++ src/main/kotlin/no/nav/sokos/skattekort.person/Application.kt
@@ -11,7 +11,7 @@ import no.nav.sokos.skattekort.person.config.commonConfig
 import no.nav.sokos.skattekort.person.config.routingConfig
 import no.nav.sokos.skattekort.person.config.securityConfig
 import no.nav.sokos.skattekort.person.database.OracleDataSource
-import no.nav.sokos.skattekort.person.gdpr.GdprLogger
+import no.nav.sokos.skattekort.person.auditlogg.AuditLogger
 import no.nav.sokos.skattekort.person.metrics.appStateReadyFalse
 import no.nav.sokos.skattekort.person.metrics.appStateRunningFalse
 import no.nav.sokos.skattekort.person.service.SkattekortPersonService
@@ -20,7 +20,7 @@ fun main() {
     val applicationState = ApplicationState()
     val applicationConfiguration = PropertiesConfig.Configuration()
     val oracleDataSource = OracleDataSource(applicationConfiguration.databaseConfig)
-    val gdprLogger = GdprLogger()
+    val gdprLogger = AuditLogger()
     val skattekortPersonService = SkattekortPersonService(oracleDataSource, gdprLogger)
 
     HttpServer(applicationState, applicationConfiguration, skattekortPersonService, oracleDataSource).start()
