+++ src/main/kotlin/no/nav/sokos/skattekort.person/auditlogg/Saksbehandler.kt
@@ -1,4 +1,4 @@
-package no.nav.sokos.skattekort.person.gdpr
+package no.nav.sokos.skattekort.person.auditlogg
 
 data class Saksbehandler(
     val ident: String
