+++ src/main/kotlin/no/nav/sokos/skattekort.person/security/TokenAuth.kt
@@ -2,7 +2,7 @@ package no.nav.sokos.skattekort.person.security
 
 import com.auth0.jwt.JWT
 import io.ktor.server.application.*
-import no.nav.sokos.skattekort.person.gdpr.Saksbehandler
+import no.nav.sokos.skattekort.person.auditlogg.Saksbehandler
 
 const val JWT_CLAIM_NAVIDENT = "NAVident"
