+++ src/main/kotlin/no/nav/sokos/skattekort.person/service/SkattekortPersonService.kt
@@ -4,21 +4,21 @@ import no.nav.sokos.skattekort.person.api.model.SkattekortPersonRequest
 import no.nav.sokos.skattekort.person.database.OracleDataSource
 import no.nav.sokos.skattekort.person.database.RepositoryExtensions.useAndHandleErrors
 import no.nav.sokos.skattekort.person.database.SkattekortPersonRepository.hentSkattekortPaaFnrOgInntektsAar
-import no.nav.sokos.skattekort.person.gdpr.Saksbehandler
+import no.nav.sokos.skattekort.person.auditlogg.Saksbehandler
 import no.nav.sokos.skattekort.person.domain.SkattekortTilArbeidsgiver
-import no.nav.sokos.skattekort.person.gdpr.GdprLogg
-import no.nav.sokos.skattekort.person.gdpr.GdprLogger
+import no.nav.sokos.skattekort.person.auditlogg.AuditLogg
+import no.nav.sokos.skattekort.person.auditlogg.AuditLogger
 
 class SkattekortPersonService(
     private val oracleDataSource: OracleDataSource,
-    private val gdrpLogger: GdprLogger
+    private val auditLogger: AuditLogger
 ) {
 
     fun hentSkattekortPerson(
         skattekortPersonRequest: SkattekortPersonRequest,
         saksbehandler: Saksbehandler
     ): List<SkattekortTilArbeidsgiver> {
-        gdrpLogger.auditLog(GdprLogg(saksbehandler = saksbehandler.ident, fnr = skattekortPersonRequest.fnr))
+        auditLogger.auditLog(AuditLogg(saksbehandler = saksbehandler.ident, fnr = skattekortPersonRequest.fnr))
         oracleDataSource.connection.useAndHandleErrors { connection ->
             return connection.hentSkattekortPaaFnrOgInntektsAar(skattekortPersonRequest)
         }
